evaluation/requirements.txt
# for evaluation
arize-phoenix[evals]
uvicorn
fastapi
# ragas
# datasets